+++ trash/inspect_mem.py
