+++ trash/isnpect2.py
