+++ trash/tablas.py
